requirements.txt
matplotlib==3.6.1
matplotlib==3.7.2
networkx==3.1
openai==0.28.1
python-decouple==3.8
torch==2.1.0
torch==2.0.1+cu117
transformers==4.35.0.dev0
websockets==11.0.3
